backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Data processing and ML dependencies
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.1
scipy==1.11.4

# Model persistence and utilities
joblib==1.3.2
python-dotenv==1.0.0

# LLM and API dependencies
requests==2.32.4

# Validation and serialization
pydantic==2.5.0
email-validator==2.1.0

# Authentication and database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
